requirements.txt
# local package
# -e .

# external requirements
Pillow
numpy

# development specific requirements
flake8
mypy
jupyter
